% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_and_count.R
\name{group_and_count}
\alias{group_and_count}
\title{group_and_count}
\usage{
group_and_count(data, ..., count_c = "cnt")
}
\arguments{
\item{data}{Input data frame}

\item{...}{Multiple value, the group column}

\item{count_c}{The count number column name, with default value cnt.}
}
\value{
A tibble with the count result
}
\description{
group_and_count
}
\details{
Group the input data with specified column and count each group's occurrence
}
\examples{
group_and_count(penguins, island)
group_and_count(penguins, island, sex)

}
